fix: move function to component
